requirements.txt
anthropic>=0.34.0
requests>=2.31.0
Pillow>=10.4.0
pytrends>=4.9.2
duckduckgo-search>=6.2.0
python-dotenv>=1.0.0
